Use red/yellow color scheme for example
... rather than red/green, which is hard to distinguish for Daltonian people.

Reasoning: When reading the docs, I first didn't understand the annotations example. It seemed that the annotation was missing, since I only saw a blue line on a uniform background. Only when looking at the code did I realize that there was a red/green pattern in the background of the plot.

Note that, according to [1]: Red–green color blindness affects up to 8% of males and 0.5% of females of Northern European descent.

[1]: https://en.wikipedia.org/wiki/Color_blindness